+++ test/Feature/CleanRegex/_prepared/template/PatternTest.php
@@ -6,7 +6,6 @@ use PHPUnit\Framework\TestCase;
 use Test\Utils\Assertion\AssertsPattern;
 use TRegx\CleanRegex\Exception\ExplicitDelimiterRequiredException;
 use TRegx\CleanRegex\Exception\MaskMalformedPatternException;
-use TRegx\CleanRegex\Exception\PlaceholderFigureException;
 use TRegx\CleanRegex\Pattern;
 use TRegx\Exception\MalformedPatternException;
 
@@ -35,34 +34,6 @@ class PatternTest extends TestCase
             'comment (but no "x" flag)'              => ["You/her #@\n her?", "%You/her #(?>X)\n her?%"],
             'comment ("x" flag, but also "-x" flag)' => ["You/her (?x:(?-x:#@\n)) her?", "%You/her (?x:(?-x:#(?>X)\n)) her?%"],
         ];
-    }
-
-    /**
-     * @test
-     */
-    public function shouldThrowForMissingTemplateAlteration()
-    {
-        // given
-        $template = Pattern::template('@@@@@');
-        // then
-        $this->expectException(PlaceholderFigureException::class);
-        $this->expectExceptionMessage('Not enough corresponding figures supplied. Used 5 placeholders, but 1 figures supplied.');
-        // when
-        $template->alteration(['foo', 'bar']);
-    }
-
-    /**
-     * @test
-     */
-    public function shouldThrowForSuperfluousTemplateAlteration()
-    {
-        // given
-        $template = Pattern::template('Foo');
-        // then
-        $this->expectException(PlaceholderFigureException::class);
-        $this->expectExceptionMessage('Supplied a superfluous figure. Used 0 placeholders, but 1 figures supplied.');
-        // when
-        $template->alteration(['foo', 'bar']);
     }
 
     /**